python/cudf/pyproject.toml
# SPDX-FileCopyrightText: Copyright (c) 2021-2026, NVIDIA CORPORATION.
# SPDX-License-Identifier: Apache-2.0

[build-system]
build-backend = "rapids_build_backend.build"
requires = [
    "rapids-build-backend>=0.4.0,<0.5.0",
    "scikit-build-core[pyproject]>=0.11.0",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../../dependencies.yaml and run `rapids-dependency-file-generator`.

[project]
name = "cudf"
dynamic = ["version"]
description = "cuDF - GPU Dataframe"
readme = { file = "README.md", content-type = "text/markdown" }
authors = [
    { name = "NVIDIA Corporation" },
]
license = "Apache-2.0"
requires-python = ">=3.11"
dependencies = [
    "cachetools",
    "cuda-python>=13.0.1,<14.0",
    "cupy-cuda13x>=13.6.0,!=14.0.0",
    "fsspec>=0.6.0",
    "libcudf==26.4.*,>=0.0.0a0",
    "numba-cuda>=0.22.2,<0.28.0",
    "numba>=0.60.0,<0.65.0",
    "numpy>=1.26,<3.0",
    "nvtx>=0.2.1",
    "packaging",
    "pandas>=2.0,<2.4.0",
    "pyarrow>=19.0.0",
    "pylibcudf==26.4.*,>=0.0.0a0",
    "rich",
    "rmm==26.4.*,>=0.0.0a0",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../../dependencies.yaml and run `rapids-dependency-file-generator`.
classifiers = [
    "Intended Audience :: Developers",
    "Topic :: Database",
    "Topic :: Scientific/Engineering",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

[project.optional-dependencies]
test = [
    "aiobotocore>=2.2.0,<3",
    "boto3>=1.21.21",
    "botocore>=1.24.21",
    "cramjam",
    "fastavro>=0.22.9",
    "hypothesis>=6.131.7",
    "msgpack",
    "pytest",
    "pytest-benchmark",
    "pytest-cases>=3.8.2",
    "pytest-cov",
    "pytest-xdist",
    "s3fs>=2022.3.0",
    "scipy",
    "tzdata",
    "zstandard",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../../dependencies.yaml and run `rapids-dependency-file-generator`.
pandas-tests = [
    "hypothesis>=6.131.7",
    "ipython",
    "pandas[test, pyarrow, performance, computation, fss, excel, parquet, feather, hdf5, spss, html, xml, plot, output-formatting, clipboard, compression]",
    "pytest-reportlog",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../../dependencies.yaml and run `rapids-dependency-file-generator`.
cudf-pandas-tests = [
    "certifi",
    "ipython",
    "jupyter_client",
    "nbconvert",
    "nbformat",
    "openpyxl",
    "pytest-rerunfailures!=16.0.0",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../../dependencies.yaml and run `rapids-dependency-file-generator`.

[project.urls]
Homepage = "https://github.com/rapidsai/cudf"
Documentation = "https://docs.rapids.ai/api/cudf/stable/"

[tool.pydistcheck]
select = [
    "distro-too-large-compressed",
]

# PyPI hard limit is 1GiB, but try to keep this as small as possible
max_allowed_size_compressed = '15M'

[tool.pytest.ini_options]
addopts = "--tb=native --strict-config --strict-markers"
empty_parameter_set_mark = "fail_at_collect"
filterwarnings = [
    "error",
    "ignore:::.*xdist.*",
    "ignore:::.*pytest.*",
    # some third-party dependencies (e.g. 'boto3') still using datetime.datetime.utcnow()
    "ignore:.*datetime.*utcnow.*scheduled for removal.*:DeprecationWarning:botocore",
    # Deprecation warning from Pyarrow Table.to_pandas() with pandas-2.2+
    "ignore:Passing a BlockManager to DataFrame is deprecated:DeprecationWarning",
    # PerformanceWarning from cupy warming up the JIT cache
    "ignore:Jitify is performing a one-time only warm-up to populate the persistent cache:cupy._util.PerformanceWarning",
    # Ignore numba PEP 456 warning specific to arm machines
    "ignore:FNV hashing is not implemented in Numba.*:UserWarning",
    # Allow running UDF tests on older architectures
    "ignore:.*NVRTC log messages whilst compiling.*:UserWarning",
]
markers = [
    "spilling: mark benchmark a good candidate to run with `CUDF_SPILL=ON`",
    "serial: Mark a test as not being parallel-safe.",
    "no_copy_on_write: A test that is not valid when copy-on-write is enabled.",
]
xfail_strict = true

[tool.rapids-build-backend]
build-backend = "scikit_build_core.build"
dependencies-file = "../../dependencies.yaml"
matrix-entry = "cuda_suffixed=true"
requires = [
    "cmake>=3.30.4",
    "cython>=3.2.2",
    "libcudf==26.4.*,>=0.0.0a0",
    "librmm==26.4.*,>=0.0.0a0",
    "ninja",
    "numba-cuda>=0.22.2,<0.28.0",
    "pylibcudf==26.4.*,>=0.0.0a0",
    "rmm==26.4.*,>=0.0.0a0",
] # This list was generated by `rapids-dependency-file-generator`. To make changes, edit ../../dependencies.yaml and run `rapids-dependency-file-generator`.

[tool.scikit-build]
build-dir = "build/{wheel_tag}"
cmake.build-type = "Release"
cmake.version = "CMakeLists.txt"
minimum-version = "build-system.requires"
ninja.make-fallback = false
sdist.exclude = ["*tests*"]
sdist.reproducible = true
wheel.packages = ["cudf"]
wheel.py-api = "cp311" # overridden in CI builds by arguments from `ci/build_wheel_*.sh` scripts.

[tool.scikit-build.metadata.version]
provider = "scikit_build_core.metadata.regex"
input = "cudf/VERSION"
regex = "(?P<value>.*)"

[tool.ruff]
extend = "../../pyproject.toml"

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["cudf"]
section-order = ["future", "standard-library", "third-party", "dask", "rapids", "first-party", "local-folder"]

[tool.ruff.lint.isort.sections]
dask = ["dask", "distributed", "dask_cuda"]
rapids = ["rmm", "pylibcudf"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]
